Create separte component for Pagination
