+++ apps/consultation-portal/components/Pagination/Pagination.css.ts
@@ -1,0 +1,12 @@
+import { style } from '@vanilla-extract/css'
+
+export const spanStyle = style({
+  position: 'absolute',
+  width: '1px',
+  height: '1px',
+  padding: '0',
+  margin: '-1px',
+  overflow: 'hidden',
+  clip: 'rect(0,0,0,0)',
+  border: '0',
+})
